fix: lowercase module filenames in signalprocessing
