+++ pylops/signalprocessing/bilinear.py
