+++ pylops/signalprocessing/dwt.py
