+++ pylops/signalprocessing/fredholm1.py
