+++ pylops/signalprocessing/seislet.py
