+++ pylops/signalprocessing/sliding2d.py
