+++ pylops/signalprocessing/sliding3d.py
@@ -3,7 +3,7 @@ from typing import List, Tuple, Union
 
 from pylops.basicoperators import BlockDiag, Diagonal, HStack, Restriction
 from pylops.LinearOperator import aslinearoperator
-from pylops.signalprocessing.Sliding2D import _slidingsteps
+from pylops.signalprocessing.sliding2d import _slidingsteps
 from pylops.utils.tapers import taper3d
 
 logging.basicConfig(format="%(levelname)s: %(message)s", level=logging.WARNING)
